+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Let `Parser` output the content of `Info.javaText` in the case of `FunctionPointer` as well
  * Fix `TokenIndexer` failure on macros using the concat `##` operator with empty arguments ([issue #525](https://github.com/bytedeco/javacpp/issues/525))
  * Let `Parser` support arrays of anonymous `struct` or `union` containing another one ([discussion #528](https://github.com/bytedeco/javacpp/discussions/528))
  * Prevent `Parser` from outputting duplicate `Pointer` constructors for basic containers